requirements.txt
# These are the main packages used to run our *.py scripts 
opencv-python
matplotlib
python-dotenv
mysqlclient
mysql
numpy
mtcnn
keras
keras-facenet
scipy
h5py
scikit-learn